refactor : 4단계 미션 진행 전 Test Code 패키지 정리
