+++ src/test/java/kitchenpos/acceptance/menu/MenuAcceptanceTest.java
@@ -1,13 +1,13 @@
-package kitchenpos.domain;
+package kitchenpos.acceptance.menu;
 
-import static kitchenpos.domain.MenuAcceptanceTestMethod.메뉴_등록_요청;
-import static kitchenpos.domain.MenuAcceptanceTestMethod.메뉴_등록되어_있음;
-import static kitchenpos.domain.MenuAcceptanceTestMethod.메뉴_등록됨;
-import static kitchenpos.domain.MenuAcceptanceTestMethod.메뉴_목록_응답됨;
-import static kitchenpos.domain.MenuAcceptanceTestMethod.메뉴_목록_조회_요청;
-import static kitchenpos.domain.MenuAcceptanceTestMethod.메뉴_목록_포함됨;
-import static kitchenpos.domain.MenuGroupAcceptanceTestMethod.메뉴_그룹_등록되어_있음;
-import static kitchenpos.domain.ProductAcceptanceTestMethod.상품_등록되어_있음;
+import static kitchenpos.acceptance.menu.MenuAcceptanceTestMethod.메뉴_등록_요청;
+import static kitchenpos.acceptance.menu.MenuAcceptanceTestMethod.메뉴_등록되어_있음;
+import static kitchenpos.acceptance.menu.MenuAcceptanceTestMethod.메뉴_등록됨;
+import static kitchenpos.acceptance.menu.MenuAcceptanceTestMethod.메뉴_목록_응답됨;
+import static kitchenpos.acceptance.menu.MenuAcceptanceTestMethod.메뉴_목록_조회_요청;
+import static kitchenpos.acceptance.menu.MenuAcceptanceTestMethod.메뉴_목록_포함됨;
+import static kitchenpos.acceptance.menugroup.MenuGroupAcceptanceTestMethod.메뉴_그룹_등록되어_있음;
+import static kitchenpos.acceptance.product.ProductAcceptanceTestMethod.상품_등록되어_있음;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
