+++ src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -5,7 +5,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 
 import java.util.List;
-import kitchenpos.application.fixture.MenuGroupFixtureFactory;
+import kitchenpos.fixture.MenuGroupFixtureFactory;
 import kitchenpos.application.menugroup.MenuGroupService;
 import kitchenpos.domain.menugroup.MenuGroup;
 import kitchenpos.domain.menugroup.MenuGroupRepository;
