+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -8,7 +8,7 @@ import static org.mockito.BDDMockito.given;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.application.fixture.ProductFixtureFactory;
+import kitchenpos.fixture.ProductFixtureFactory;
 import kitchenpos.application.product.ProductService;
 import kitchenpos.domain.product.Product;
 import kitchenpos.domain.product.ProductRepository;
