+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -7,14 +7,14 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.application.fixture.MenuFixtureFactory;
-import kitchenpos.application.fixture.MenuGroupFixtureFactory;
-import kitchenpos.application.fixture.MenuProductFixtureFactory;
-import kitchenpos.application.fixture.OrderFixtureFactory;
-import kitchenpos.application.fixture.OrderLineItemFixtureFactory;
-import kitchenpos.application.fixture.OrderTableFixtureFactory;
-import kitchenpos.application.fixture.ProductFixtureFactory;
-import kitchenpos.application.fixture.TableGroupFixtureFactory;
+import kitchenpos.fixture.MenuFixtureFactory;
+import kitchenpos.fixture.MenuGroupFixtureFactory;
+import kitchenpos.fixture.MenuProductFixtureFactory;
+import kitchenpos.fixture.OrderFixtureFactory;
+import kitchenpos.fixture.OrderLineItemFixtureFactory;
+import kitchenpos.fixture.OrderTableFixtureFactory;
+import kitchenpos.fixture.ProductFixtureFactory;
+import kitchenpos.fixture.TableGroupFixtureFactory;
 import kitchenpos.application.order.OrderService;
 import kitchenpos.application.table.TableService;
 import kitchenpos.domain.menu.Menu;
