+++ src/test/java/kitchenpos/domain/order/OrderLineItemsTest.java
@@ -6,9 +6,9 @@ import static org.assertj.core.api.Assertions.assertThatNoException;
 import com.google.common.collect.Lists;
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.application.fixture.MenuFixtureFactory;
-import kitchenpos.application.fixture.MenuGroupFixtureFactory;
-import kitchenpos.application.fixture.OrderTableFixtureFactory;
+import kitchenpos.fixture.MenuFixtureFactory;
+import kitchenpos.fixture.MenuGroupFixtureFactory;
+import kitchenpos.fixture.OrderTableFixtureFactory;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menugroup.MenuGroup;
 import kitchenpos.domain.table.OrderTable;
