+++ src/test/java/kitchenpos/fixture/MenuProductFixtureFactory.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.fixture;
+package kitchenpos.fixture;
 
 
 import kitchenpos.domain.menu.Menu;
