+++ src/test/java/kitchenpos/fixture/OrderFixtureFactory.java
@@ -1,11 +1,10 @@
-package kitchenpos.application.fixture;
+package kitchenpos.fixture;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.table.OrderTable;
 
 public class OrderFixtureFactory {
